requirements.txt
pandas>=2.2
pyarrow>=14.0
rich>=13.7
click>=8.1
beautifulsoup4>=4.12
lxml>=5.2
PyYAML>=6.0
